+++ HW05/HW05/Rectangle2D.hpp
@@ -22,23 +22,23 @@ public:
     
     // Constructors
     Rectangle2D();
-    Rectangle2D(int x1, int y1, int width1, int height1);
+    Rectangle2D(double x1, double y1, double width1, double height1);
     
     // Getters and setters for x and y values
-    const int getX();
-    const int getY();
-    int setX(int newX);
-    int setY(int newY);
+    const double getX();
+    const double getY();
+    double setX(double newX);
+    double setY(double newY);
     
     // Getters and setters for the width and height
-    const int getWidth();
-    const int getHeight();
-    int setWidth(int newWidth);
-    int setHeight(int newHeight);
+    const double getWidth();
+    const double getHeight();
+    double setWidth(double newWidth);
+    double setHeight(double newHeight);
     
     // Functions to calculate area and perimeter
-    const int getArea();
-    const int getPerimeter();
+    const double getArea();
+    const double getPerimeter();
     
     // Functions to check relationships between two rectangles
     const bool contains(double x, double y);
